camera_test: Adds option to choose target board.
[Comment]
This patch adds option to provide target board as one
of the argument.
This is added because Tablet-1 and Tablet-2 secondary
sensors supports 30fps but not blaze.
To make the hundreads of test scripts reusable across
the targets, this patch overwrites the fps to 30 if
the script passes 27fps for tablet-1 and tablet-2
targets.

[DR]
-
Signed-off-by: Solaiyappan Saravanan <saravanan.s@ti.com>

Signed-off-by: Jean Johnson <a0271255@ti.com>
Change-Id: I8b6e32c4f974a6c1d4fde9a4a1fae0e84cda10c2