CameraHAL: Fix for first time face detection.
There is a issue in face detection in case of sensor mount orientation
is not zero and the onOrientationEvent() is not triggered.

In omap5, the sensor mount orientation is != 0. But the startFaceDetection()
sets the orienation which is available in mDeviceOrientation. This
mDeviceOrientation is initialized to zero and only gets updated in
OnOrientationEvent() where the sensor mount orientation also considered.

Because of this the face detection will not find any faces until the
onOrientationEvent() get called.

This patch initialize the mDeviceOrientation to sensor mount orientation.
So that faceDetection will find face starts from the preview even before
the board moved around to get the onOrientationEvent().

Change-Id: Iecf912b309e7551ed10cbae4ba32be860d0a9beb
Signed-off-by: Solaiyappan Saravanan <saravanan.s@ti.com>